Compute empirical population by imputing reponses and measured sizes using knn.

Compute empirical population by imputing reponses and measured sizes using knn.

Usage

get_empirical_population(sample_indices, pop, y)

Arguments

sample_indices

Indices of sample.

pop

Population data frame to be sampled with 4 columns.

Halton numbers

X1-coordinate of population unit

X2-coordinate of population unit

Size measurement of population unit

y

Sample response values.

Value

A summary data frame of the estimator.